backend/requirements.txt
alembic==1.12.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.7.0
click==8.1.7
flasgger==0.9.7b2
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
Mako==1.3.0
MarkupSafe==2.1.3
mistune==3.0.2
mysqlclient==2.2.0
packaging==23.2
pathlib==1.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.31.0
rpds-py==0.13.1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
